+++ src/components/MainDrawer.vue
@@ -2,14 +2,20 @@
 import TabSearchVue from "./TabSearch.vue";
 import TabSettingsVue from "./TabSettings.vue";
 
-import { inject, Ref } from "vue";
-import { IConfig } from "./../types/config";
-import { Status, map } from "../utils/status";
+import { ref, shallowReactive, toRaw, inject, provide, watch, Ref, ShallowReactive } from "vue";
+import { VueI18nTranslation } from "vue-i18n";
+import { Notification } from "@arco-design/web-vue";
 
-const visible = inject("visible") as Ref<boolean>; // 是否显示
+import { IConfig } from "./../types/config";
+import { INotebooks, Data_fullTextSearchBlock } from "./../types/siyuan";
+import { Status, map } from "./../utils/status";
+import { Method, SiyuanClient } from "./../utils/siyuan";
+
 const status = inject("status") as Ref<Status>; // 连接状态
 const message = inject("message") as Ref<string>; // 连接状态消息
 const version = inject("version") as Ref<string>; // 内核版本
+
+const visible = inject("visible") as Ref<boolean>; // 是否显示
 
 const config = inject("config") as IConfig; // 用户配置
 
@@ -19,6 +25,69 @@ function handleOk() {
 function handleCancel() {
     visible.value = false;
 }
+
+/* 👇 查询内容 👇 */
+const query = ref(""); // 查询语句
+const keywords = ref<string[]>([]); // 查询关键词
+const results = shallowReactive<Data_fullTextSearchBlock>({
+    blocks: [],
+    matchedBlockCount: 0,
+    matchedRootCount: 0,
+}); // 查询结果
+
+/* 将关键字列表转换为查询语句 */
+function keywords2query(value: string[]): string {
+    return value.join(" ");
+}
+
+/* 查询关键词 => 查询语句 */
+watch(keywords, value => {
+    query.value = keywords2query(value);
+});
+
+const client = inject("client") as InstanceType<typeof SiyuanClient>; // 客户端
+const notebooks = inject("notebooks") as ShallowReactive<INotebooks>; // 笔记本列表
+
+/* 搜索 */
+async function search($t: VueI18nTranslation, keyword: boolean) {
+    try {
+        const q = keyword ? keywords2query(keywords.value) : query.value;
+        if (q.length === 0) {
+            results.blocks = [];
+            results.matchedBlockCount = 0;
+            results.matchedRootCount = 0;
+            return;
+        }
+
+        if (notebooks.list.length === 0) {
+            const response = await client.lsNotebooks();
+            notebooks.list = response.data.notebooks;
+        }
+
+        // REF [响应式 API：进阶 toRay() | Vue.js](https://cn.vuejs.org/api/reactivity-advanced.html#toraw)
+        // REF [响应式 API：工具函数 unref() | Vue.js](https://cn.vuejs.org/api/reactivity-utilities.html#unref)
+        const payload = Object.assign({}, toRaw(config.search), {
+            query: q,
+        });
+
+        const response = await client.fullTextSearchBlock(payload);
+
+        results.blocks = response.data.blocks;
+        results.matchedBlockCount = response.data.matchedRootCount;
+        results.matchedRootCount = response.data.matchedBlockCount;
+    } catch (error) {
+        console.warn(error);
+        Notification.error({
+            title: $t("search"),
+            content: String(error),
+            closable: true,
+            duration: 3000,
+        });
+    }
+}
+
+provide("results", results);
+/* 👆 查询内容 👆 */
 </script>
 
 <template>
@@ -90,9 +159,21 @@ function handleCancel() {
                 <!-- REF [Arco Design Vue](https://arco.design/vue/component/input-tag) -->
                 <!-- 搜索输入框 -->
                 <a-input-tag
+                    v-if="config.search.method === Method.keyword"
+                    class="title-input-tag"
+                    size="mini"
+                    v-model:model-value="keywords"
+                    :placeholder="$t('search')"
+                    @change="search($t, true)"
+                    allow-clear
+                />
+                <a-input
+                    v-else
                     class="title-input"
                     size="mini"
+                    v-model="query"
                     :placeholder="$t('search')"
+                    @change="search($t, false)"
                     allow-clear
                 />
             </div>
@@ -114,6 +195,10 @@ function handleCancel() {
 <style lang="less">
 // 支持 .arco-* 选择器需要移除 scoped 标签
 #siyuan-drawer {
+    .arco-drawer-title {
+        width: 100%;
+    }
+
     .title {
         display: flex;
         align-items: center;
@@ -130,7 +215,9 @@ function handleCancel() {
             font-size: inherit;
             flex: none;
         }
-        .title-input {
+
+        .title-input,
+        .title-input-tag {
             flex: auto;
         }
     }